+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/debugger/breakpoint/RobotLineBreakpointType.java
@@ -1,24 +1,24 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.ide.debugger.breakpoint;
 
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotFeatureFileType;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotLanguage;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotResourceFileType;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotTokenTypes;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordDefinitionId;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordInvokable;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.VariableDefinition;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Ref;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiWhiteSpace;
-import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.xdebugger.XDebuggerUtil;
 import com.intellij.xdebugger.breakpoints.SuspendPolicy;
 import com.intellij.xdebugger.breakpoints.XLineBreakpointType;
 import com.jetbrains.python.debugger.PyDebugSupportUtils;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotFeatureFileType;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotLanguage;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotResourceFileType;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotTokenTypes;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Heading;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordDefinition;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordDefinitionId;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordStatement;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.VariableDefinition;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -62,13 +62,9 @@ public class RobotLineBreakpointType extends XLineBreakpointType<RobotLineBreakp
     }
 
     @SuppressWarnings("SuspiciousMethodCalls")
-    protected void lineHasStoppablePsi(@NotNull Project project,
-                                       int line,
-                                       Document document,
-                                       Ref<? super Boolean> stoppable) {
+    protected void lineHasStoppablePsi(@NotNull Project project, int line, Document document, Ref<? super Boolean> stoppable) {
         XDebuggerUtil.getInstance().iterateLine(project, document, line, psiElement -> {
-            if (psiElement.getNode() != null &&
-                       Set.of(RobotTokenTypes.WHITESPACE, RobotTokenTypes.ERROR).contains(psiElement.getNode().getElementType())) {
+            if (psiElement.getNode() != null && Set.of(RobotTokenTypes.WHITESPACE, RobotTokenTypes.ERROR).contains(psiElement.getNode().getElementType())) {
                 return true;
             } else {
                 if (isPsiElementStoppable(psiElement)) {
@@ -84,17 +80,10 @@ public class RobotLineBreakpointType extends XLineBreakpointType<RobotLineBreakp
 
     protected boolean isPsiElementStoppable(PsiElement psiElement) {
         if (psiElement.getLanguage() == RobotLanguage.INSTANCE) {
-            if (!(psiElement instanceof PsiWhiteSpace) && PsiTreeUtil.getChildOfType(psiElement.getParent(), KeywordInvokable.class) != null) {
-                return true;
-            }
-            PsiElement nextSibling = psiElement.getNextSibling();
-            if (PsiTreeUtil.getChildOfType(nextSibling, KeywordInvokable.class) != null) {
-                return true;
-            }
-            if (nextSibling instanceof VariableDefinition && PsiTreeUtil.findChildOfType(nextSibling, KeywordInvokable.class) != null) {
-                return true;
-            }
-            return !(psiElement instanceof PsiWhiteSpace) && psiElement.getParent() instanceof KeywordDefinitionId;
+            PsiElement context = psiElement.getContext();
+            // TODO: Not optimal solution. Better use visitor pattern here, but the visitor needs to be implemented first
+            return context instanceof KeywordStatement || context instanceof VariableDefinition || context instanceof KeywordDefinition
+                   || context instanceof KeywordDefinitionId || context instanceof Heading heading && heading.isSettings();
         }
         return false;
     }